write which file failed when they fail

+++ test/prog/run_all.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-set -e
+failed=0
 
 for xl_file in test/prog/*.xl; do
     if grep skip-run $xl_file >/dev/null; then
@@ -8,7 +8,17 @@ for xl_file in test/prog/*.xl; do
         continue
     fi
     dist/expelc $xl_file a.out
+    if [ "$?" != "0" ]; then
+        failed=1
+        echo "fail: $xl_file in compilation"
+        continue
+    fi
     dist/runexpel a.out >/dev/null
-    rm a.out
-    echo "ok:   $xl_file"
+    if [ "$?" != "0" ]; then
+        failed=1
+        echo "fail: $xl_file"
+    else
+        rm a.out
+        echo "ok:   $xl_file"
+    fi
 done
